bugfix on last day of tournament
Nantzbot and website updater both use calculateCurrentTournamentId to identify the current tournament. But, website updater writes files, and the calculateCurrentTournamentId depends in part on counting files. So Nantzbot sees that the last written file is in tournament finished state, and assumes that it is the first day of the next tournament. But that is wrong-there is no file yet for that day; Add a boolean to tell calculateCurrentTournamentId whether the file has yet been written for today